Add ability to insert configs for comprtconfigs
Also, debcomprt now manages the comprtconfigs repo slightly by noticing
if the repo is already on the filesystem to just try and pull for
updates. Though, this has not been fully tested yet. Filemode has been looked
into for the progDataDir and it seems my inital implementation was
incorrect.

Smaller refactors also occured.