Change filemodes for mount directories
There does not seem to be a definite specification for what the mount
directories permissions should be, though there seem to exist a few
pointers of recommendations.

Ideally, the principle of least privilege seems to apply here with how
my current workstation has permissions set for the mount directories.

The difference being was write permissions for 'user' is omitted from /proc
and /sys. Otherwise I would have stayed with the 755 filemode.