Work on TODOs, refactored code
Public/Available make stuff has been changed to common. As really both
seem suggest a false reality. In that, not all available make targets
will be shown and there is no such thing as 'public' anything for make.

debsource is now just deb. As the binary package is also built.