TestChrootCommandIntegration has been added
That said, a lot of other changes have been added as well. I will highlight
those that were needed to add the new integration test.

Prior to this commit, if a comprt was not created from an alias, then
no default user was added unless a comprtconfig had knowledge about
this.
I decided if an alias was opted for, then a default user (using the same
uid) would be added to the comprt.

I would like to try and refactor the tests file further but for now
setting up the progDataDir has been placed in a function. As both
integration tests did make use of the logic.
This needs to eventually makes its way into
TestCreateCommandIntegration.