Work on TODOs
Subcommands should be implemented now to be non-optional. In that, if
debcomprt is called with no subcommands. The help menu will print and
exit with a status code of 1.

In the event any part of the chrooting fails, now the process of exiting
will attempt to unmount any filesystems that were mounted prior.

In the case of the chroot shell session returning anything but a 0
exit status. I decided I'd rather error on the side of caution and just
allow go to panic and display the error code and/or message. Trying to
determine each actual error would be pretty much impossible todo.