Do DISCUSS on cmdArgs now progConfigs
I think the idea I was going with here is that the type name does not
fully reveal everything that this type is used for. Hence the name change.

It was discovered that the recent addition of having -a and -c arguments
being incompatible was not implemented correctly. Hence fixed.