Implement DISCUSS on function signatures
Generating the debootstrap arguments was seen as something outside of
creating a comprt. Thus, I decided to create an entirely new function for
doing that task and generating the arguments outside of creating a
comprt.
The major benefit here was avoiding a larger function signature.

Another option thought about was a function for generating progConfigs.
This would generating a struct with 'sane' defaults. However that
route does not seem benefical at the time.