Integrate docker container for building deb pkg
It was discovered that during packaging, that the packaging was somewhat
tied to the host OS that the package was created on. To mitigate this,
I decided to create a docker container that would act as a target debian
env to perform packaging in.

It was also explored of using pbuilder for creating an env that would do
that same thing. However, it would be more time consuming or manually create
a new env each time during the creation of the debian packaging.

That, and Docker containers are the more 'modern' and versatile.