Make it clearer that debcomprt must run as root
I notice when coming back to this project, that debcomprt must run as
root or else make will fail to run when preprocessing aliases.

That said, the error given was not clear and was not obvious that not running
as root was the issue. Thus a fix has been added to make this more
clear.